svgtoipe: Fix parsing elliptical arcs with minimal whitespace
svgtoipe.py crashed on android.svg from the following webpage: https://uxwing.com/android-studio-icon/

	Traceback (most recent call last):
	  File "/opt/ipe-tools/svgtoipe/svgtoipe.py", line 1164, in <module>
	    main()
	  File "/opt/ipe-tools/svgtoipe/svgtoipe.py", line 1159, in main
	    svg.parse_svg(args.outfile)
	  File "/opt/ipe-tools/svgtoipe/svgtoipe.py", line 479, in parse_svg
	    self.write_data()
	  File "/opt/ipe-tools/svgtoipe/svgtoipe.py", line 496, in write_data
	    self.parse_nodes(self.root)
	  File "/opt/ipe-tools/svgtoipe/svgtoipe.py", line 511, in parse_nodes
	    getattr(self, nodeName)(n)
	  File "/opt/ipe-tools/svgtoipe/svgtoipe.py", line 751, in node_g
	    self.parse_nodes(group)
	  File "/opt/ipe-tools/svgtoipe/svgtoipe.py", line 511, in parse_nodes
	    getattr(self, nodeName)(n)
	  File "/opt/ipe-tools/svgtoipe/svgtoipe.py", line 1100, in node_path
	    parse_path(self.out, d)
	  File "/opt/ipe-tools/svgtoipe/svgtoipe.py", line 263, in parse_path
	    rx, ry, phi, large_arc, sweep, x2, y2 = pnext(d, 7)
						    ^^^^^^^^^^^
	  File "/opt/ipe-tools/svgtoipe/svgtoipe.py", line 176, in pnext
	    l.append(float(d.pop(0)))
		     ^^^^^^^^^^^^^^^
	ValueError: could not convert string to float: 'l'

I believe the offending bit of svg path data is this:

	a2.507 2.507 0 002.68-2.352 2.503 2.503 0 00-2.348-2.68l-24.46-1.748

Note the "002.68" which is two '0' flags followed by the number '2.68'.

I considered changing the regex to disallow leading zeros,
but that's not consistent with the spec, and that also wouldn't work in
case the flags are '1' instead of '0'.

I added a link to the editor's draft version of the SVG spec
instead of the official spec at w3.org because the one currently hosted
on w3.org has a bad EBNF grammar for "number".